+++ R/confint.R
@@ -162,38 +162,10 @@ confint.focus_list <- function(object,
 }
 
 .confint_profile_focus_list_glm <- function(object, level, nleqslv_args) {
-    fit <- object$object
-    if (!identical(fit$type, "ML")) {
-        fit <- update(fit, type = "ML", start = coef(fit, model = "mean"))
-    }
-    p_mean <- length(coef(fit, model = "mean"))
-    theta <- coef(fit, model = "full")
-    if (fit$family$family %in% c("poisson", "binomial")) {
-        theta <- theta[names(coef(fit, model = "mean"))]
-    }
-    aux <- enrichwith::get_auxiliary_functions(fit)
-    m_inds <- seq_len(p_mean)
-    d_ind <- p_mean + 1
-    split_theta <- function(theta) {
-        if (length(theta) == p_mean) {
-            list(coefficients = theta)
-        } else {
-            list(coefficients = theta[m_inds],
-                 dispersion = theta[d_ind])
-        }
-    }
-    loglik <- function(theta)
-        sum(do.call(aux$dmodel, c(split_theta(theta), list(log = TRUE))))
-    score <- function(theta)
-        do.call(aux$score, split_theta(theta))
-    information <- function(theta)
-        do.call(aux$information, split_theta(theta))
+    components <- .profile_glm_components(object)
     do.call(profile_ci,
-            c(list(loglik = loglik,
-                   score = score,
-                   information = information,
-                   mle = theta,
-                   on = object$on$on,
+            c(components,
+              list(on = object$on$on,
                    on_gradient = object$on$on_gradient,
                    on_hessian = object$on$on_hessian,
                    level = level,